move sql storage implementation into a separate submodule
