+++ mlos_bench/config/storage/duckdb.jsonc
@@ -1,6 +1,6 @@
 // Local DuckDB storage.
 {
-    "class": "mlos_bench.storage.sql_storage.SqlStorage",
+    "class": "mlos_bench.storage.sql.storage.SqlStorage",
 
     // Apply --config-path resolution to these config parameters:
     "resolve_path": ["database"],
