+++ mlos_bench/config/storage/postgresql.jsonc
@@ -1,6 +1,6 @@
 // PostgreSQL storage.
 {
-    "class": "mlos_bench.storage.sql_storage.SqlStorage",
+    "class": "mlos_bench.storage.sql.storage.SqlStorage",
 
     "config": {
         "log_sql": false,  // Write all SQL statements to the log.
