+++ mlos_bench/config/storage/sqlite.jsonc
@@ -1,6 +1,6 @@
 // Local SQLite3 database.
 {
-    "class": "mlos_bench.storage.sql_storage.SqlStorage",
+    "class": "mlos_bench.storage.sql.storage.SqlStorage",
 
     // Apply --config-path resolution to these config parameters:
     "resolve_path": ["database"],
