+++ mlos_bench/mlos_bench/storage/sql/experiment.py
@@ -15,8 +15,8 @@ from sqlalchemy import Engine, Connection, Table, column, func
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
 from mlos_bench.storage.base_storage import Storage
-from mlos_bench.storage.sql_schema import DbSchema
-from mlos_bench.storage.sql_trial import Trial
+from mlos_bench.storage.sql.schema import DbSchema
+from mlos_bench.storage.sql.trial import Trial
 
 _LOG = logging.getLogger(__name__)
