+++ mlos_bench/mlos_bench/storage/sql/storage.py
@@ -14,8 +14,8 @@ from sqlalchemy import URL, create_engine
 from mlos_bench.tunables.tunable_groups import TunableGroups
 from mlos_bench.service.base_service import Service
 from mlos_bench.storage.base_storage import Storage
-from mlos_bench.storage.sql_schema import DbSchema
-from mlos_bench.storage.sql_experiment import Experiment
+from mlos_bench.storage.sql.schema import DbSchema
+from mlos_bench.storage.sql.experiment import Experiment
 
 _LOG = logging.getLogger(__name__)
